Include arm neon sources in CMake build.
The CMake build sets PNG_ARM_NEON_OPT to 2 if __ARM_NEON__ is detected, but it wasn't using the arm source files required for ARM neon support to work.

This fixes that by always including the files and relying on the preprocessor and PNG_ARM_NEON_OPT to activate the source code for ARM builds only.